requirements.txt
requests
streamlit
python-dotenv
langchain
langchain-openai
pydantic
firecrawl-py
